SECURITY UPDATE: Use POST method for PowerDNS communication
This update changes the internal PowerDNS communication to use the POST
method. This prevents random websites from injecting domain names into
the programmatic DNS server.

Previously, the GET method was used. However, this was a security risk
since the GET method on a domain was actually non-idempotent, in that it
allowed any website to affect the cookie IPv6 mapping (see
https://website.peterjin.org/wiki/Transparent_Happy_Eyeballs)

We are still vulnerable to DNS rebinding. However, it does not matter in
our application since the programmatic DNS server can only return
"cookie" IPv6 addresses for Internet domains. This concern may be
reconsidered for "split DNS" use cases where only some domains use the
programmatic DNS server, and other domains use a regular DNS server to
connect to the regular Internet.