index: Add 'wild' mode
'Wild' mode is intended for a Socketbox setup. The idea is that instead
of restricting the connections to a single prefix, wild mode allows any
prefix to be used. However, there is no translation with respect to the
programmatic DNS server, so this can only be used with relay_map, SNI
proxy, and the SOCKS and PP2 servers.